refactor: router agent structure
